Fix image path validation position on duplicates
Previously when using a text search based approach to calculating the
position for a diagnostic, if the same broken image was present more
than once in a document, all of the diagnostics would point to the
location of the first. This fixes the issue by using a more reliable
approach based upon the line and column data returned in the xpath
results for elements.