Add filtering support for ToC trees view
This change allows the ability to filter on title or the string used
in the description field (e.g. for modules the ID) using the built-in
capabilities of the VScode tree widget by clicking on the filter
icon. This fully expands the tree, so a collapse-all button is also
exposed to allow users to close the tree.