Shamefully add sleeps to failing test cases
These sleep additions are the only way I was able to find to fix
tests from failing with errors of the following form when updating
to versions of vscode newer than 1.55.2:

    Error: Could NOT open editor for "file:///...".

More details including a simplified test case that reproduces the
error can be found in https://github.com/openstax/cnx/issues/1569.
This hack can hopefully be removed in the future, but I'm going this
route after being unable to root cause and find an alternative fix in
a timebox period.