Fix caching issue with module imageSources
This change addresses an issue where the memoize-one cache isn't
busted when changes occur to bundle media files (e.g. renaming image
files) resulting in missing problem diagnostics. Since the caching
needs to remain at the module level to be effective, it changes the
method signature to require a provided cached bundleMedia set that is
used to populate properties in the generated ImageSource data.