Move push content validation into _pushContent (#144)
* Move push content validation into _pushContent

After a commit message is supplied, open modified documents to
let the xml extension find any schema errors that may exist. If
there are errors, stop the push operation.

* Update push content tests

Since openAndValidate was added to _pushContent, it causes
problems with the tests involving that function. Stub it
in the first round of pushContent tests since it is tested in depth later.

* Make canPush sync because it does not need to be async