Make server send onDidChangeWatchedFiles to client again
Previously, the client listened for a onDidChangeWatchedFiles and used
it to drive many updates. Since then, the ToC was updated to use a
different mechanism for updates.

As far as I can tell, the only thing that still uses this request to
drive updates is the CNXML preview panel

The TocEditorPanel references onDidChangeWatchedFiles; however, there
is a second mechanism that causes it to update to match the server's
ToC. This reference to onDidChangeWatchedFiles may be redundant.