When moving node to different book, update both collection files
I was stuck on this for a bit. Initially I tried
to write the srcBookToc immediately after removing the node on
model-manager.ts:594. The tests kept failing because the page was being
orphaned instead of appearing in the second book.

I realized that the sideEffectFn was running which
overwrote `this.bookTocs` with a new array. This meant that the splice
(model-manager.ts:597) was modifying a different list of children because the
reference returned by `lookupToken` was not the same as the one stored in
`bookToc`.

I moved my write operation to the end and added a warning. Hopefully the
warning will save someone else from this confusion in the future.