Track orphan loading with promise instead of bool
This does not actually change coverage hits, so the bool was probably working as intended. I think this approach is much easier to reason with.

It's more clear that the intent is to run this once and only once. It also lessens the need to understand the lack of side-effects caused by repeated calls in the methods that used to be called repeatedly.

I also added a timeout. That will become important soon.