Autocomplete should only load orphans when needed
Instead of assuming that they are needed for all autocomplete handlers, let the handlers themselves handle it.

Add a timeout to orphan loading to unblock the autocomplete handler after half a second. It seems like autocomplete for all extensions hangs if one handler awaits a long task.

I have not verified this, but I suspect vscode is doing something like `Promise.all` to gather results which would explain the behavior nicely.

This does have the downside of making early autocomplete results incomplete. Until everything is finished loading, the orphans are effectively isolated to the currently opened document.