Core 1032 ancillary schema (#218)
* Track orphan loading with promise instead of bool

This does not actually change coverage hits, so the bool was probably working as intended. I think this approach is much easier to reason with.

It's more clear that the intent is to run this once and only once. It also lessens the need to understand the lack of side-effects caused by repeated calls in the methods that used to be called repeatedly.

I also added a timeout. That will become important soon.

* Autocomplete should only load orphans when needed

Instead of assuming that they are needed for all autocomplete handlers, let the handlers themselves handle it.

Add a timeout to orphan loading to unblock the autocomplete handler after half a second. It seems like autocomplete for all extensions hangs if one handler awaits a long task.

I have not verified this, but I suspect vscode is doing something like `Promise.all` to gather results which would explain the behavior nicely.

This does have the downside of making early autocomplete results incomplete. Until everything is finished loading, the orphans are effectively isolated to the currently opened document.

* Remove manual schema modifications

These are not required anymore since we are going a different direction and they conflict with the generation process.

* Add md:tags and md:tag to schema

For integration with ancillary service.

* Add mdml to xml catalog

Not sure why it was missing. This enables autocomplete, among other things, via the xml extension.

* Add author; nest author and tags inside super

The nesting the relationship and purpose more clear

* Change author to subject name

* Re-enable ancillary creation; remove unused `slug`

* Add validation for super metadata

Only valid when document class is "super"

* Minor fixes

Squish `expectValue` into value instead of type assertion

Improve condition for page metadata validation

* Remove link from tag schema